pyproject.toml
[project]
name = "topnotch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "acres==0.5.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "ci-info==0.3.0",
    "click==8.3.0",
    "cohere==5.20.0",
    "colorama==0.4.6",
    "configobj==5.0.9",
    "configparser==7.2.0",
    "etelemetry==0.3.1",
    "fastapi>=0.121.1",
    "fastavro==1.12.1",
    "filelock==3.20.0",
    "fsspec==2025.10.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==1.1.2",
    "idna==3.11",
    "looseversion==1.3.0",
    "lxml==6.0.2",
    "mcp>=1.0.0",
    "networkx==3.5",
    "nibabel==5.3.2",
    "nipype==1.10.0",
    "numpy==2.3.4",
    "openai>=2.7.2",
    "openai-agents>=0.5.0",
    "packaging==24.2",
    "pandas==2.3.3",
    "pathlib==1.0.1",
    "pinecone>=5.0.0",
    "prov==2.1.1",
    "puremagic==1.30",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydot==4.0.1",
    "pymupdf>=1.26.6",
    "pyparsing==3.2.5",
    "pypdf2==3.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyxnat==1.6.4",
    "pyyaml==6.0.3",
    "rdflib==7.4.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "scipy==1.16.3",
    "shellingham==1.5.4",
    "simplejson==3.20.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tqdm==4.67.1",
    "traits==7.0.2",
    "typer-slim==0.20.0",
    "types-requests==2.32.4.20250913",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uv==0.9.8",
    "uvicorn>=0.38.0",
]
